Fix : 이벤트중복 등록 오류 수정
